Do not override Tezos endpoint
Problem: in bridge.nix, we override chains.tezos.custom.endpoint
field of the config to point to our testnet node.
I don't know why it was done in the first place, but I think it's
not needed anymore because config.yaml in bridge-web already points
to our testnet node.

Solution: remove it from `config`.